+++ execdriver/docker/default_template.go
@@ -1,4 +1,4 @@
-package namespaces
+package docker
 
 import (
 	"github.com/dotcloud/docker/pkg/cgroups"
